+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ fluent_compiler 0.4 (under development)
 
 * Dropped support for Python 2.7 and Python < 3.6
 * Lots of code style cleanups using black etc.
+* Compiler actually respects documented ``use_isolating`` attribute of escapers
 
 fluent_compiler 0.3 (2020-11-18)
 --------------------------------